+++ etna/analysis/feature_relevance/relevance_table.py
@@ -1,5 +1,3 @@
-from abc import ABC, abstractmethod
-
 import numpy as np
 import pandas as pd
 
@@ -33,18 +31,3 @@ def get_statistics_relevance_table(df: pd.DataFrame, df_exog: pd.DataFrame) -> p
     relevance_table.index = segments
     relevance_table.columns = regressors
     return relevance_table
-
-
-class RelevanceTable(ABC):
-    def __init__(self, greater_is_better: bool):
-        self.greater_is_better = greater_is_better
-
-    @abstractmethod
-    def __call__(self, *args, **kwargs) -> pd.DataFrame:
-        pass
-
-
-class StatisticsRelevanceTable(RelevanceTable):
-    def __call__(self, df: pd.DataFrame, df_exog: pd.DataFrame) -> pd.DataFrame:
-        table = get_statistics_relevance_table(df=df, df_exog=df_exog)
-        return table